+++ src/settings/MappingProfiles/MappingProfilesTransformationsModal/MappingProfilesTransformationsModal.js
@@ -22,7 +22,11 @@ import {
   CollapseFilterPaneButton,
 } from '@folio/stripes/smart-components';
 
-import { recordTypes } from '../RecordTypeField';
+import {
+  RECORD_TYPES,
+  SELECTED_STATUS,
+  SELECTED_STATUSES,
+} from '../../../utils';
 import { TransformationsForm } from './TransformationsForm';
 import { SearchForm } from './SearchForm';
 import { normalizeTransformationFormValues } from './TransformationsField';
@@ -31,7 +35,15 @@ import css from './MappingProfilesTransformationsModal.css';
 
 const initialSearchFormValues = {
   searchValue: '',
-  filters: { recordTypes: recordTypes.map(recordType => recordType.value) },
+  filters: {
+    recordTypes: RECORD_TYPES.map(({ value }) => value),
+    statuses: SELECTED_STATUSES.map(({ value }) => value),
+  },
+};
+
+const statusesFilterMap = {
+  [SELECTED_STATUS.SELECTED]: ({ fieldId }, selectedTransformations) => Boolean(selectedTransformations[fieldId]),
+  [SELECTED_STATUS.UNSELECTED]: ({ fieldId }, selectedTransformations) => !selectedTransformations[fieldId],
 };
 
 const fullWidthStyle = { style: { width: '100%' } };
@@ -69,11 +81,23 @@ export const MappingProfilesTransformationsModal = ({
   const searchResults = [];
   let displayedCheckedItemsAmount = 0;
 
+  const filterMap = {
+    recordTypes: (filters, transformation) => filters.includes(transformation.recordType),
+    statuses: (filters, transformation) => filters.some(status => statusesFilterMap[status](transformation, selectedTransformations)),
+  };
+
   searchValueResults.forEach(transformation => {
-    if (get(searchFilters, 'recordTypes', []).includes(transformation.recordType)) {
+    const hasFilterMatch = Object.keys(filterMap)
+      .every(filterKey => {
+        const filter = get(searchFilters, filterKey, []);
+
+        return filterMap[filterKey](filter, transformation);
+      });
+
+    if (hasFilterMatch) {
       searchResults.push(transformation);
 
-      if (selectedTransformations[transformation.order]) {
+      if (selectedTransformations[transformation.fieldId]) {
         displayedCheckedItemsAmount++;
       }
     }